test: move key manager helper to test-utils
